+++ source/js/account/refreshToken.ts
@@ -1,19 +1,19 @@
+import {sendMessageToBackground} from '../tools/sendMessageToBackground';
+import {triggerPromise} from '../tools/triggerPromise';
+import {getErrorAsString} from '../tools/getErrorMessage';
+import {GroupResponder} from '../tools/GroupResponder';
+import {StateChange} from '../messaging/MessageType';
+import {warn} from '../log/log';
 import {readSession} from './readSession';
 import {fetchJson, jsonRequest} from '../api';
 import {saveSession} from './saveSession';
 import type {Session} from './Session';
-import {sendMessageToBackground} from '../tools/sendMessageToBackground';
-import {StateChange} from '../messaging/MessageType';
-import {triggerPromise} from '../tools/triggerPromise';
-import {warn} from '../log/log';
 import {RefreshTokenError} from './RefreshTokenError';
-import {getErrorAsString} from '../tools/getErrorMessage';
 
-let refreshing = false;
-let waitingRefreshPromises = [] as [
-	(session: Session) => void,
-	(error: Error | RefreshTokenError | string) => void,
-][];
+const tokenRequestsGroup = new GroupResponder<
+	Session,
+	Error | RefreshTokenError | string
+>();
 
 export const hasRefreshToken = async (session?: Session): Promise<boolean> => {
 	if (!session) {
@@ -28,80 +28,80 @@ export const refreshToken = async (session?: Session): Promise<Session> => {
 		session = await readSession();
 	}
 
-	if (!session.uid || !session.refreshToken) {
+	const uid = session.uid;
+
+	if (!uid || !session.refreshToken) {
 		throw new RefreshTokenError();
 	}
 
-	if (refreshing) {
-		return await new Promise((resolve, reject) => {
-			waitingRefreshPromises.push([resolve, reject]);
-		});
-	}
+	return tokenRequestsGroup.handle(async () => {
+		const now = Date.now();
 
-	try {
-		refreshing = true;
-		session.expiresAt = Date.now();
-		await saveSession(session);
-		const auth = await fetchJson<{
-			RefreshToken: string;
-			AccessToken: string;
-		}>(
-			'auth/refresh',
-			jsonRequest(
-				'POST',
-				{
-					UID: session.uid,
-					ResponseType: 'token',
-					GrantType: 'refresh_token',
-					RefreshToken: session.refreshToken,
-					RedirectURI: session.redirectURI || 'https://protonvpn.com',
-				},
-				{
-					'x-pm-uid': session.uid,
-				},
-			),
-		);
-		delete session.expiresAt;
-		session.refreshToken = auth.RefreshToken;
-		session.accessToken = auth.AccessToken;
-		await saveSession(session);
+		try {
+			session.expiresAt = now;
+			await saveSession(session);
+			const auth = await fetchJson<{
+				RefreshToken: string;
+				AccessToken: string;
+			}>(
+				'auth/refresh',
+				jsonRequest(
+					'POST',
+					{
+						UID: uid,
+						ResponseType: 'token',
+						GrantType: 'refresh_token',
+						RefreshToken: session.refreshToken,
+						RedirectURI: session.redirectURI || 'https://protonvpn.com',
+					},
+					{
+						'x-pm-uid': uid,
+					},
+				),
+			);
 
-		waitingRefreshPromises.forEach(([resolve]) => {
-			resolve(session as Session);
-		});
-		waitingRefreshPromises = [];
+			const sessionAfterApiCall = await readSession();
 
-		return session;
-	} catch (e) {
-		let error = e;
-		const nextSession = await readSession();
-
-		if (
-			nextSession.expiresAt &&
-			nextSession.accessToken === session.accessToken
-		) {
-			delete nextSession.expiresAt;
-			await saveSession(nextSession);
-		}
-
-		if ((error as any).Code === 10013) {
-			warn(error, new Error().stack);
-			error = new RefreshTokenError(getErrorAsString(error));
-
-			try {
-				triggerPromise(sendMessageToBackground(StateChange.SIGN_OUT));
-			} catch {
-				// Let the initial error throw
+			// Modified by something else in the meantime, or emptied by logout
+			if (sessionAfterApiCall.expiresAt !== now) {
+				return sessionAfterApiCall;
 			}
 
-			waitingRefreshPromises.forEach(([, reject]) => {
-				reject(error as RefreshTokenError);
-			});
-			waitingRefreshPromises = [];
-		}
+			delete session.expiresAt;
+			session.refreshToken = auth.RefreshToken;
+			session.accessToken = auth.AccessToken;
+			await saveSession(session);
 
-		throw error;
-	} finally {
-		refreshing = false;
-	}
+			return session;
+		} catch (e) {
+			let error = e;
+			const nextSession = await readSession();
+
+			// Modified by something else in the meantime, or emptied by logout
+			if (nextSession.expiresAt !== now) {
+				return nextSession;
+			}
+
+			if (
+				nextSession.expiresAt &&
+				nextSession.accessToken === session.accessToken
+			) {
+				delete nextSession.expiresAt;
+				await saveSession(nextSession);
+			}
+
+			if ((error as any).Code === 10013) {
+				warn(error, new Error().stack);
+				error = new RefreshTokenError(getErrorAsString(error));
+
+				try {
+					triggerPromise(sendMessageToBackground(StateChange.SIGN_OUT));
+				} catch {
+					// Let the initial error throw
+				}
+			}
+
+			throw error;
+		}
+	});
 };