+++ source/js/transmit.ts
@@ -1,17 +1,4 @@
 import {getRuntime} from './tools/getRuntime';
+import {relayMessagesToExtension} from './tools/relayMessagesToExtension';
 
-((runtime) => {
-	runtime.connect();
-
-	window.addEventListener(
-		'message',
-		(event) => {
-			if (event.source != window) {
-				return;
-			}
-
-			runtime.sendMessage(event.data);
-		},
-		false,
-	);
-})(getRuntime());
+relayMessagesToExtension(window, getRuntime());